app/requirements.txt
# requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Werkzeug==2.3.6
SQLAlchemy==2.0.9
WTForms==3.0.1
email-validator==2.0.0
Pillow==9.5.0
python-dotenv==1.0.0
gunicorn==20.1.0